refactor(files): extract transfer engine; drop stat-before-read (short segment = EOF)
