+++ AGENTS.md
@@ -17,7 +17,7 @@ TypeScript SDK for Railway sandbox create/exec/destroy.
 - Unit tests must stay offline and must not call Railway.
 - Public API: `Sandbox` IS the sandbox via static factories — `import { Sandbox } from "railway"`, then `Sandbox.create()` / `Sandbox.connect(id)` / `Sandbox.list()`. The constructor is private; never `new Sandbox(...)`.
 - `token`/`environmentId`/`endpoint` resolve explicit → env (`RAILWAY_API_TOKEN`/`RAILWAY_ENVIRONMENT_ID`/`RAILWAY_GRAPHQL_ENDPOINT`) → default; missing credentials throw `RailwayAuthError`.
-- Keep public SDK scope minimal: `Sandbox.create`, `exec`/`fork`/`destroy`/`refresh`/`checkpoint` on the instance, `connect`/`list`/`checkpoints`/`renameCheckpoint`/`deleteCheckpoint` statics, `await using` auto-destroy. `fork` reuses the `sandboxCreate` mutation via `sourceSandboxId` (`create(source)` is the static form). Checkpoints: `checkpoint(name)` captures a running sandbox's disk (`sandboxCheckpointCreate`, synchronous — bootable when it resolves); `Sandbox.create(name)` boots from it via `input.template.name`; `key` is the user-given name (or recipe hash for built templates); rename/delete are by id; duplicate names error. Recipe builds (`Sandbox.template()`) poll the `sandboxTemplateBuild` query to READY. Files: `sandbox.files` (`read`/`write`/`list`/`stat`/`exists`/`mkdir`/`remove`/`rename`) speaks the `/ws/files` endpoint — one connection per operation, authorized by a `generateShellToken` JWT with `scope: "files:read"` (`"files:read files:write"` for mutations); content moves as 64KB binary frames, `write` declares size only when known (0 = unknown/uncapped, so unknown-length streams work), and reads/writes speak Web Streams so massive files stay memory-bounded. `write`'s `mode` option: the server does not yet apply the protocol's mode on write, so the SDK applies it with a chmod over the exec primitive after the upload (the mode is still sent in `write_start` so it picks up native support automatically; drop the chmod then). Two server timing constraints shape the client (see comments in `src/core/files-ws-client.ts` / `src/sandbox/files.ts`): uploads keep the WS send buffer shallow (128KB high-water) and downloads are issued as adaptive bounded-range segments, because sessions are killed when keepalive replies lag or a single read request streams too long. Ports and exec stdin are future.
+- Keep public SDK scope minimal: `Sandbox.create`, `exec`/`fork`/`destroy`/`refresh`/`checkpoint` on the instance, `connect`/`list`/`checkpoints`/`renameCheckpoint`/`deleteCheckpoint` statics, `await using` auto-destroy. `fork` reuses the `sandboxCreate` mutation via `sourceSandboxId` (`create(source)` is the static form). Checkpoints: `checkpoint(name)` captures a running sandbox's disk (`sandboxCheckpointCreate`, synchronous — bootable when it resolves); `Sandbox.create(name)` boots from it via `input.template.name`; `key` is the user-given name (or recipe hash for built templates); rename/delete are by id; duplicate names error. Recipe builds (`Sandbox.template()`) poll the `sandboxTemplateBuild` query to READY. Files: `sandbox.files` (`read`/`write`/`list`/`stat`/`exists`/`mkdir`/`remove`/`rename`) speaks the `/ws/files` endpoint — one connection per operation, authorized by a `generateShellToken` JWT with `scope: "files:read"` (`"files:read files:write"` for mutations); content moves as 64KB binary frames, `write` declares size only when known (0 = unknown/uncapped, so unknown-length streams work), and reads/writes speak Web Streams so massive files stay memory-bounded. `write`'s `mode` option: the server does not yet apply the protocol's mode on write, so the SDK applies it with a chmod over the exec primitive after the upload (the mode is still sent in `write_start` so it picks up native support automatically; drop the chmod then). Two server timing constraints shape the client (see comments in `src/core/files-ws-client.ts` / `src/sandbox/files-transfer.ts`): uploads keep the WS send buffer shallow (128KB high-water) and downloads are issued as adaptive bounded-range segments resumed by byte position, because sessions are killed when keepalive replies lag or a single read request streams too long. Reads never stat first (a short segment signals EOF); only `fromEnd` needs the size. The transfer engine (segmentation, resume, upload-source normalization, retry predicates) lives in `src/sandbox/files-transfer.ts`; `src/sandbox/files.ts` is the public surface. Ports and exec stdin are future.
 - Release: pushing a `v*.*.*` tag triggers `release.yml`, which publishes the `package.json` version **at the tagged commit** to npm. Invariant: `package.json` must already equal the tag's version in the commit the tag points to, or `npm publish` fails with "cannot publish over previously published versions".
   - Main way (manual): on `main`, `npm version <x.y.z> --no-git-tag-version`, `pnpm install --lockfile-only`, commit the bump, push `main`, then `git tag v<x.y.z>` on that commit and push the tag.
   - The `Create Release` workflow (workflow_dispatch) is currently broken — checkout fails because the `RELEASE_TOKEN` repo secret is missing. Don't use it until the secret is fixed.
